reorder stats.ts, extract iterateOverTrace.ts
